backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
fastapi-users[sqlalchemy]==12.1.3
pyotp==2.9.0
qrcode==7.4.2

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2

# Task Queue
celery[redis]==5.3.4
redis==5.0.1
flower==2.0.1

# Timezone
pytz==2023.3

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
fastapi-cors==0.0.6

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Scraping support
beautifulsoup4==4.12.2
lxml==5.0.0